backend/central_server/requirements.txt
fastapi>=0.115.12
uvicorn[standard]==0.24.0
httpx>=0.27.0
pydantic>=2.8.0
openai>=1.35.0
python-dotenv
numpy==1.26.4 # numpy 추가